fix(frontend): make dashboard export-safe for static build
